+++ CHANGELOG.md
@@ -7,6 +7,9 @@ publication of Payroll Information as fname `earnings-statement-for-all`.
 
 ### (Unreleased)
 
++ fix: log at DEBUG level the SOAP message when encountering a SOAP fault in the HRS integration (
+  [HRSPLT-408][], [#172][])
+
 ### 6.0.5 avoid popups in surplus earnings statement hyperlinks
 
 2018-12-10
@@ -855,6 +858,7 @@ This and many more earlier releases exist as [releases in the GitHub repo][].
 [#167]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/167
 [#168]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/168
 [#169]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/169
+[#172]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/172
 
 [HRSPLT-346]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-346
 [HRSPLT-348]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-348
@@ -883,3 +887,4 @@ This and many more earlier releases exist as [releases in the GitHub repo][].
 [HRSPLT-402]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-402
 [HRSPLT-403]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-403
 [HRSPLT-404]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-404
+[HRSPLT-404]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-408